+++ EscapeApp/app/src/androidTest/java/com/context/kroket/escapeapp/minigames/A_CodeCrackerCodeviewTest.java
@@ -1,96 +1,149 @@
 package com.context.kroket.escapeapp.minigames;
 
-import android.support.test.espresso.ViewAssertion;
-import android.support.test.espresso.assertion.ViewAssertions;
-import android.support.test.espresso.intent.rule.IntentsTestRule;
-import android.support.test.espresso.matcher.ViewMatchers;
-import android.support.test.filters.LargeTest;
-import android.support.test.runner.AndroidJUnit4;
+import android.test.ActivityInstrumentationTestCase2;
+import android.view.View;
+import android.widget.Button;
+import android.widget.EditText;
+import android.widget.TextView;
 
 import com.context.kroket.escapeapp.R;
 import com.context.kroket.escapeapp.mainscreens.MainActivity;
+import com.context.kroket.escapeapp.mainscreens.WaitingActivity;
+import com.robotium.solo.Solo;
 
-import org.junit.Rule;
+import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import static android.support.test.espresso.Espresso.onView;
-import static android.support.test.espresso.action.ViewActions.click;
-import static android.support.test.espresso.action.ViewActions.typeText;
-import static android.support.test.espresso.assertion.ViewAssertions.*;
-import static android.support.test.espresso.intent.Intents.intended;
-import static android.support.test.espresso.intent.matcher.ComponentNameMatchers.hasShortClassName;
-import static android.support.test.espresso.intent.matcher.IntentMatchers.hasComponent;
-import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
-import static android.support.test.espresso.matcher.ViewMatchers.withHint;
-import static android.support.test.espresso.matcher.ViewMatchers.withId;
-import static android.support.test.espresso.assertion.ViewAssertions.matches;
-import static android.support.test.espresso.matcher.ViewMatchers.withText;
-import static junit.framework.Assert.assertEquals;
-import static org.hamcrest.CoreMatchers.not;
 
 /**
- * Created by Harvey van Veltom on 19/05/2016.
+ * Class tests the A_CodeCrackerCodeView class. mainly the verify button and message.
+ * test class uses robotium instead of espresso because we need to close android's
+ * keyboard.
  */
-@RunWith(AndroidJUnit4.class)
-@LargeTest
-/**
- * A_CodeCrackerCodeviewTest is used to test the A_CodeCrackerCodeview class.
- */
-public class A_CodeCrackerCodeviewTest {
+public class A_CodeCrackerCodeviewTest  extends
+        ActivityInstrumentationTestCase2<MainActivity> {
 
+    /**
+     * Solo is used by robotium to interact with the game's ui elements.
+     */
+    private Solo solo;
 
-    @Rule
-    public IntentsTestRule<MainActivity> myActivityRule = new IntentsTestRule<MainActivity>(MainActivity.class);
+    /**
+     * Constructor for the test class.
+     * Requirement of robotium.
+     */
+    public A_CodeCrackerCodeviewTest() {
+        super(MainActivity.class);
+    }
+
+    /**
+     * Set's up solo object that is used by robotium
+     * to interact with ui elements.
+     *
+     * @throws Exception
+     */
+    public void setUp() throws Exception {
+        solo = new Solo(getInstrumentation(), getActivity());
+    }
+
 
 
     /**
      * Method checks the functionality of A_CodeCrackerCodeview pic button.
      */
     @Test
-    public void clickPicButton() {
-        //Switch from MainActivity to A_CodeCrackerCodeview.
-        MainActivity.switchGameAA(true);
-        onView(ViewMatchers.withId(R.id.connectButton)).perform(click());
+    public void testclickPicButton() {
+        MainActivity.TestActivtiy = MainActivity.ActivitySwitch.acode;
+        Button buttonConnect = (Button) solo.getView(R.id.connectButton);
+        solo.clickOnView(buttonConnect);
 
-        //Click the pic button.
-        onView(withId(R.id.picButton)).perform(click());
+        Button picButton = (Button) solo.getView(R.id.picButton);
+        solo.clickOnView(picButton);
 
-        //Check if we instantiated a A_Code_Cracker_Pictureview.
-        intended(hasComponent(hasShortClassName(".minigames.A_Code_Cracker_Pictureview")));
-        MainActivity.switchGameAA(false);
+        solo.assertCurrentActivity("should be pictureView", A_Code_Cracker_Pictureview.class);
+        MainActivity.TestActivtiy = MainActivity.ActivitySwitch.notest;
+
     }
 
-//    @Test
-//    public void testWrongAnswer() {
-//        //Switch from MainActivity to A_CodeCrackerCodeview.
-//        MainActivity.switchGameAA(true);
-//        onView(ViewMatchers.withId(R.id.connectButton)).perform(click());
-//
-//        //Click the pic button.
-//        onView(withId(R.id.verifyButton)).perform(click());
-//
-//        assertEquals(A_CodeCrackerCodeview.getNumberOfAttempts(), 1);
-//        //onView(withId(R.id.verifyMessage)).check(matches( isDisplayed()));
-//        MainActivity.switchGameAA(false);
-//    }
-//
-//    @Test
-//    public void testRightAnswer() {
-//        //Switch from MainActivity to A_CodeCrackerCodeview.
-//        MainActivity.switchGameAA(true);
-//        A_CodeCrackerCodeview.enableTesting(true);
-//        onView(ViewMatchers.withId(R.id.connectButton)).perform(click());
-//
-//        onView(withId(R.id.answerA)).perform(typeText("2719173"));
-//        //Click the pic button.
-//
-//
-//        onView(withId(R.id.verifyButton)).perform(click());
-//
-//        onView(withId(R.id.waiting)).check(matches( isDisplayed()));
-//        MainActivity.switchGameAA(false);
-//        A_CodeCrackerCodeview.enableTesting(false);
-//    }
+    /**
+     * The method checks to see what happens when we enter the wrong
+     * answer in A_CodeCrackerCodeView.
+     */
+    @Test
+    public void testWrongAnswer() {
+        MainActivity.TestActivtiy = MainActivity.ActivitySwitch.acode;
+        Button buttonConnect = (Button) solo.getView(R.id.connectButton);
+        solo.clickOnView(buttonConnect);
 
+        //Hide the android keyboard.
+        solo.hideSoftKeyboard();
+
+        //Click the verify button of A_CodeCrackerCodeView when
+        //we don't have an answer filled in.
+        Button buttonVerify = (Button) solo.getView(R.id.verifyButton);
+        solo.clickOnView(buttonVerify);
+
+        //Check if the verifyMessage view is visible(only becomes visible)
+        //if we have filled in a wrong answer.
+        //TextView verify = (TextView) solo.getView(R.id.verifyMessage);
+        //assertEquals("view should be visible", verify.getVisibility(), View.VISIBLE);
+
+        try {
+            Thread.sleep(1000);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+
+        assertEquals(A_CodeCrackerCodeview.getNumberOfAttempts(), 1);
+        MainActivity.TestActivtiy = MainActivity.ActivitySwitch.notest;
+
+    }
+
+    /**
+     * The method checks to see what happens when we enter the right
+     * answer in A_CodeCrackerCodeView.
+     */
+    @Test
+    public void testRightAnswer() {
+        //Switch from MainActivity to A_CodeCrackerCodeView.
+        MainActivity.TestActivtiy = MainActivity.ActivitySwitch.acode;
+        A_CodeCrackerCodeview.enableTesting(true);
+        Button buttonConnect = (Button) solo.getView(R.id.connectButton);
+        solo.clickOnView(buttonConnect);
+
+        //Fill in the correct answer in the answerA view.
+        EditText answer = (EditText) solo.getView(R.id.answerA);
+        solo.clearEditText(answer);
+        solo.typeText(answer, "2719173");
+
+        //Hide the android keyboard.
+        solo.hideSoftKeyboard();
+
+        //Temporarily switch of serviceIsBound.
+        //A_CodeCrackerCodeview.setServiceIsBound(false);
+
+        //Click the verify button of A_CodeCrackerCodeView when
+        //we have the right answer filled in.
+        Button buttonVerify = (Button) solo.getView(R.id.verifyButton);
+        solo.clickOnView(buttonVerify);
+
+        //Assert we are now in the WaitingActivity class.
+        solo.assertCurrentActivity("should be waiting", WaitingActivity.class);
+
+        MainActivity.TestActivtiy = MainActivity.ActivitySwitch.notest;
+        A_CodeCrackerCodeview.enableTesting(false);
+
+        // A_CodeCrackerCodeview.setServiceIsBound(true);
+    }
+
+
+    /**
+     * Close all activities of solo at the end of a test.
+     *
+     * @throws Exception
+     */
+    @Override
+    public void tearDown() throws Exception {
+        solo.finishOpenedActivities();
+    }
 }
